Set up email and added  roles and permissions

+++ routes/web.php
@@ -14,42 +14,39 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [VoteController::class,'index'])->name('home');
 Route::post('/', [VoteController::class,'store'])->name('vote.store');
 
-Route::middleware(['auth', 'verified'])->group(function (){
+Route::middleware(['role:admin|staff', 'verified'])->group(function (){
     Route::get('admin/dashboard', [DashboardController::class, 'index'])->name('dashboard');
-    // Route::resource('admin/election', ElectionController::class);
+
     Route::get('admin/election', [ElectionController::class, 'index'])->name('election.index');
-    Route::post('admin/election', [ElectionController::class, 'store'])->name('election.store');
-    Route::put('admin/election/{election}', [ElectionController::class, 'update'])->name('election.update');
-    Route::delete('admin/election/{election}', [ElectionController::class, 'destroy'])->name('election.destroy');
+    Route::post('admin/election', [ElectionController::class, 'store'])->name('election.store')->middleware('permission:can create');
+    Route::put('admin/election/{election}', [ElectionController::class, 'update'])->name('election.update')->middleware('permission:can update');
+    Route::delete('admin/election/{election}', [ElectionController::class, 'destroy'])->name('election.destroy')->middleware('permission:can delete');
 
-    // Route::resource('admin/position', PositionController::class);
     Route::get('admin/position', [PositionController::class, 'index'])->name('position.index');
-    Route::post('admin/position', [PositionController::class,'store'])->name('position.store');
-    Route::put('admin/position/{position}', [PositionController::class, 'update'])->name('position.update');
-    Route::delete('admin/position/{position}', [PositionController::class, 'destroy'])->name('position.destroy');
+    Route::post('admin/position', [PositionController::class,'store'])->name('position.store')->middleware('permission:can create');
+    Route::put('admin/position/{position}', [PositionController::class, 'update'])->name('position.update')->middleware('permission:can update');
+    Route::delete('admin/position/{position}', [PositionController::class, 'destroy'])->name('position.destroy')->middleware('permission:can delete');
 
-    // Route::resource('admin/candidate', CandidateController::class);
     Route::get('admin/candidate', [CandidateController::class, 'index'])->name('candidate.index');
-    Route::post('admin/candidate', [CandidateController::class,'store'])->name('candidate.store');
-    Route::put('admin/candidate/{candidate}', [CandidateController::class, 'update'])->name('candidate.update');
-    Route::delete('admin/candidate/{candidate}', [CandidateController::class, 'destroy'])->name('candidate.destroy');
+    Route::post('admin/candidate', [CandidateController::class,'store'])->name('candidate.store')->middleware('permission:can create');
+    Route::put('admin/candidate/{candidate}', [CandidateController::class, 'update'])->name('candidate.update')->middleware('permission:can update');
+    Route::delete('admin/candidate/{candidate}', [CandidateController::class, 'destroy'])->name('candidate.destroy')->middleware('permission:can delete');
 
-    // Route::resource('admin/voter', VoterController::class);
     Route::get('admin/voter', [VoterController::class, 'index'])->name('voter.index');
-    Route::post('admin/voter', [VoterController::class,'store'])->name('voter.store');
+    Route::post('admin/voter', [VoterController::class,'store'])->name('voter.store')->middleware('permission:can create');
     Route::put('admin/voter/{voter}', [VoterController::class, 'update'])->name('voter.update');
-    Route::delete('admin/voter/{voter}', [VoterController::class, 'destroy'])->name('voter.destroy');
+    Route::delete('admin/voter/{voter}', [VoterController::class, 'destroy'])->name('voter.destroy')->middleware('permission:can delete');
     Route::post('admin/voter/upload', [VoterController::class,'upload'])->name('voter.upload');
-    // Route::resource('admin/votes', VoteController::class);
     
     Route::get('admin/archive', [ArchiveController::class, 'index'])->name('archive.index');
     Route::get('admin/archive/{archive}', [ArchiveController::class, 'show'])->name('archive.show');
+});
 
-    // Route::resource('admin/user', UserController::class);
+Route::middleware(['role:admin', 'verified'])->group(function (){
     Route::get('admin/user', [UserController::class, 'index'])->name('user.index');
-    Route::post('admin/user', [UserController::class,'store'])->name('user.store');
-    Route::put('admin/user/{user}', [UserController::class, 'update'])->name('user.update');
-    Route::delete('admin/user/{user}', [UserController::class, 'destroy'])->name('user.destroy');
+    Route::post('admin/user', [UserController::class,'store'])->name('user.store')->middleware('permission:can create');
+    Route::put('admin/user/{user}', [UserController::class, 'update'])->name('user.update')->middleware('permission:can update');
+    Route::delete('admin/user/{user}', [UserController::class, 'destroy'])->name('user.destroy')->middleware('permission:can delete');
 });
 
 Route::middleware('auth')->group(function () {
